% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.command.with.argument.R
\name{add.command.with.argument}
\alias{add.command.with.argument}
\title{add.command.with.argument}
\usage{
add.command.with.argument(sbd, command, argument)
}
\description{
add.command.with.argument
}

