% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.command.with.autonumber.R
\name{add.command.with.autonumber}
\alias{add.command.with.autonumber}
\title{add.command.with.autonumber}
\usage{
add.command.with.autonumber(sbd, command, autonumber, ...)
}
\description{
add.command.with.autonumber
}
