% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.command.with.nolabel.R
\name{add.command.with.nolabel}
\alias{add.command.with.nolabel}
\title{add.command.with.nolabel}
\usage{
add.command.with.nolabel(sbd, command, ...)
}
\description{
add.command.with.nolabel
}

