% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.correlation.matrix.R
\name{extract.correlation.matrix}
\alias{extract.correlation.matrix}
\title{extract.correlation.matrix}
\usage{
extract.correlation.matrix(file, path = "")
}
\description{
extract.correlation.matrix
}

