% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.covariance.matrix.R
\name{extract.covariance.matrix}
\alias{extract.covariance.matrix}
\title{$Id: extract.covariance.matrix.R 1828 2007-11-30 01:16:07Z adunn $}
\usage{
extract.covariance.matrix(file, path = "")
}
\description{
$Id: extract.covariance.matrix.R 1828 2007-11-30 01:16:07Z adunn $
}

