% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.free.parameters.R
\name{extract.free.parameters}
\alias{extract.free.parameters}
\title{extract.free.parameters}
\usage{
extract.free.parameters(file, path = "", type = c("values", "bounds",
  "both"))
}
\description{
extract.free.parameters
}

