% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.initial.state.R
\name{extract.initial.state}
\alias{extract.initial.state}
\title{Get initial state from output file}
\usage{
extract.initial.state(file, path)
}
\description{
This requires that \code{initial_state} is set to \code{T} in the \code{output.sbd} file.
}

