% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.objective.function.R
\name{extract.objective.function}
\alias{extract.objective.function}
\title{extract.objective.function}
\usage{
extract.objective.function(file, path = "")
}
\description{
extract.objective.function
}

