% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.profiles.R
\name{extract.profiles}
\alias{extract.profiles}
\title{Extracts data from a casal -p run}
\usage{
extract.profiles(filename)
}
\description{
Outputs a list with one component for each parameter that is profiled
Each component is a list with components
values = n x 2 matrix with columns Value (parameter value) and
         Objective (objective function at that value)
         for n points in profile
header = string containing description of parameter vector
parameters = n x npar matrix contining parameter values at each profile point
}
\author{
R.I.C.C. Francis
}

