% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.projection.R
\name{extract.projection}
\alias{extract.projection}
\title{Extract projection results to data frame}
\usage{
extract.projection(file = "", path = "", write.to.file = TRUE)
}
\description{
Extract projection results to data frame
}
