% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.quantities.actural_catches.R
\name{extract.quantities.actural_catches}
\alias{extract.quantities.actural_catches}
\title{extract.quantities.actural_catches}
\usage{
extract.quantities.actural_catches(lines)
}
\description{
extract.quantities.actural_catches
}
