% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.quantities.fishing_pressures.R
\name{extract.quantities.fishing_pressures}
\alias{extract.quantities.fishing_pressures}
\title{extract.quantities.fishing_pressures}
\usage{
extract.quantities.fishing_pressures(lines)
}
\description{
extract.quantities.fishing_pressures
}
