% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.quantities.from.table.R
\name{extract.quantities.from.table}
\alias{extract.quantities.from.table}
\title{$Id: extract.quantities.from.table.R 1828 2007-11-30 01:16:07Z adunn $}
\usage{
extract.quantities.from.table(file, path = "")
}
\description{
$Id: extract.quantities.from.table.R 1828 2007-11-30 01:16:07Z adunn $
}

