% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.quantities.lambda.R
\name{extract.quantities.lambda}
\alias{extract.quantities.lambda}
\title{extract.quantities.lambda}
\usage{
extract.quantities.lambda(lines)
}
\description{
extract.quantities.lambda
}

