% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.quantities.mapped.initialisation_parameters.R
\name{extract.quantities.mapped.initialisation_parameters}
\alias{extract.quantities.mapped.initialisation_parameters}
\title{extract.quantities.mapped.initialisation_parameters}
\usage{
extract.quantities.mapped.initialisation_parameters(lines)
}
\description{
extract.quantities.mapped.initialisation_parameters
}

