% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/extract.quantities.mapped.resight_p.parameters.R
\name{extract.quantities.mapped.resight_p.parameters}
\alias{extract.quantities.mapped.resight_p.parameters}
\title{extract.quantities.mapped.resight_p.parameters}
\usage{
extract.quantities.mapped.resight_p.parameters(lines)
}
\description{
extract.quantities.mapped.resight_p.parameters
}
