% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.quantities.mapped.survival.parameters.R
\name{extract.quantities.mapped.survival.parameters}
\alias{extract.quantities.mapped.survival.parameters}
\title{extract.quantities.mapped.survival.parameters}
\usage{
extract.quantities.mapped.survival.parameters(lines)
}
\description{
extract.quantities.mapped.survival.parameters
}

