% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.quantities.mapped.transition.parameters.R
\name{extract.quantities.mapped.transition.parameters}
\alias{extract.quantities.mapped.transition.parameters}
\title{extract.quantities.mapped.transition.parameters}
\usage{
extract.quantities.mapped.transition.parameters(lines)
}
\description{
extract.quantities.mapped.transition.parameters
}

