% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.quantities.mark_recapture.R
\name{extract.quantities.mark_recapture}
\alias{extract.quantities.mark_recapture}
\title{extract.quantities.mark_recapture}
\usage{
extract.quantities.mark_recapture(temp)
}
\description{
extract.quantities.mark_recapture
}

