% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.quantities.scalar.parameter.values.R
\name{extract.quantities.scalar.parameter.values}
\alias{extract.quantities.scalar.parameter.values}
\title{extract.quantities.scalar.parameter.values}
\usage{
extract.quantities.scalar.parameter.values(lines)
}
\description{
extract.quantities.scalar.parameter.values
}

