% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.quantities.selectivity_at_class.R
\name{extract.quantities.selectivity_at_class}
\alias{extract.quantities.selectivity_at_class}
\title{extract.quantities.selectivity_at_class}
\usage{
extract.quantities.selectivity_at_class(lines)
}
\description{
extract.quantities.selectivity_at_class
}

