% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.sbd.file.R
\name{extract.sbd.file}
\alias{extract.sbd.file}
\title{$Id: extract.csl.file.R 1830 2007-11-30 01:37:18Z adunn $}
\usage{
extract.sbd.file(file, path = "")
}
\description{
$Id: extract.csl.file.R 1830 2007-11-30 01:37:18Z adunn $
}

