% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.sbd.command.R
\name{get.sbd.command}
\alias{get.sbd.command}
\title{$Id: get.sbd.command.R 1750 2007-11-13 00:45:01Z adunn $}
\usage{
get.sbd.command(object, command, label, where.subcommand)
}
\description{
$Id: get.sbd.command.R 1750 2007-11-13 00:45:01Z adunn $
}

