% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.populate.tmatrix.R
\name{populate}
\alias{populate}
\alias{populate.tmatrix}
\title{Transition matrix constructor function}
\usage{
populate(x, ...)

\method{populate}{tmatrix}(object, from, to, name, verbose = TRUE)
}
\description{
Transition matrix constructor function
}

