% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sbd.commands.R
\docType{data}
\name{sbd.commands}
\alias{sbd.commands}
\title{$Id: sbd.commands.R 1750 2007-11-13 00:45:01Z adunn $}
\format{\preformatted{List of 2
 $ command: chr [1:47] "comment" "comment" "comment" "initial" ...
 $ type   : chr [1:47] "argument" "argument" "argument" "argument" ...
}}
\usage{
sbd.commands
}
\description{
$Id: sbd.commands.R 1750 2007-11-13 00:45:01Z adunn $
}
\keyword{datasets}

