% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.convert.lines.to.matrix.R
\name{SeaBird.convert.lines.to.matrix}
\alias{SeaBird.convert.lines.to.matrix}
\title{SeaBird.convert.lines.to.matrix}
\usage{
SeaBird.convert.lines.to.matrix(lines, from = -1, to = -1, contains = "",
  starts.with = "", clip.to = "", clip.from = "", clip.to.match = "",
  clip.from.match = "", row.names = FALSE, ...)
}
\description{
SeaBird.convert.lines.to.matrix
}

