% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.convert.to.lines.R
\name{SeaBird.convert.to.lines}
\alias{SeaBird.convert.to.lines}
\title{$Id: SeaBird.convert.to.lines.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.convert.to.lines(filename)
}
\description{
$Id: SeaBird.convert.to.lines.R 1747 2007-11-12 21:01:38Z adunn $
}

