% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.get.classes.R
\name{SeaBird.get.classes}
\alias{SeaBird.get.classes}
\title{Get vector of class names from \code{population.sbd}}
\usage{
SeaBird.get.classes(filename)
}
\description{
Get vector of class names from \code{population.sbd}
}

