% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.get.lines.R
\name{SeaBird.get.lines}
\alias{SeaBird.get.lines}
\title{$Id: SeaBird.get.lines.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.get.lines(lines, from = -1, to = -1, contains = "",
  starts.with = "", clip.to = "", clip.from = "", clip.to.match = "",
  clip.from.match = "", ...)
}
\description{
$Id: SeaBird.get.lines.R 1747 2007-11-12 21:01:38Z adunn $
}

