% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.is.whole.number.R
\name{SeaBird.is.whole.number}
\alias{SeaBird.is.whole.number}
\title{$Id: SeaBird.is.whole.number.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.is.whole.number(string)
}
\description{
$Id: SeaBird.is.whole.number.R 1747 2007-11-12 21:01:38Z adunn $
}

