% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.make.filename.R
\name{SeaBird.make.filename}
\alias{SeaBird.make.filename}
\title{$Id: SeaBird.make.filename.R 1834 2007-12-02 21:32:26Z adunn $}
\usage{
SeaBird.make.filename(file, path = "")
}
\description{
$Id: SeaBird.make.filename.R 1834 2007-12-02 21:32:26Z adunn $
}

