% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.make.list.R
\name{SeaBird.make.list}
\alias{SeaBird.make.list}
\title{$Id: SeaBird.make.list.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.make.list(lines)
}
\description{
$Id: SeaBird.make.list.R 1747 2007-11-12 21:01:38Z adunn $
}

