% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.make.sbd.commands.R
\name{SeaBird.make.sbd.commands}
\alias{SeaBird.make.sbd.commands}
\title{$Id: SeaBird.make.sbd.commands.R 1801 2007-11-22 21:15:33Z adunn $}
\usage{
SeaBird.make.sbd.commands(file = "dictionary.cpp", path)
}
\description{
$Id: SeaBird.make.sbd.commands.R 1801 2007-11-22 21:15:33Z adunn $
}

