% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.make.table.R
\name{SeaBird.make.table}
\alias{SeaBird.make.table}
\title{$Id: SeaBird.make.table.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.make.table(lines)
}
\description{
$Id: SeaBird.make.table.R 1747 2007-11-12 21:01:38Z adunn $
}

