% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.pos.match.R
\name{SeaBird.pos.match}
\alias{SeaBird.pos.match}
\title{$Id: SeaBird.pos.match.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.pos.match(vector, regexp)
}
\description{
$Id: SeaBird.pos.match.R 1747 2007-11-12 21:01:38Z adunn $
}

