% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.pos.R
\name{SeaBird.pos}
\alias{SeaBird.pos}
\title{$Id: SeaBird.pos.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.pos(vector, x)
}
\description{
$Id: SeaBird.pos.R 1747 2007-11-12 21:01:38Z adunn $
}

