% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.R
\name{SeaBird}
\alias{SeaBird}
\title{Run SeaBird from within an R terminal session}
\usage{
SeaBird(exe.dir = ".", sbd.dir = exe.dir, results.dir = exe.dir,
  external = FALSE)
}
\arguments{
\item{exe.dir}{location of SeaBird.exe}

\item{sbd.dir}{location of input files}

\item{results.dir}{location in which to save mpd output}

\item{external}{logical indicating whether an external command window should be opened}
}
\description{
Currently only runs an MPD fit using \code{SeaBird -e}. The session output is piped to a file called \code{mpd.out}.
}

