% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.recodevector.R
\name{SeaBird.recodevector}
\alias{SeaBird.recodevector}
\title{$Id: SeaBird.recodevector.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.recodevector(in.data, from.vals, to.vals)
}
\description{
$Id: SeaBird.recodevector.R 1747 2007-11-12 21:01:38Z adunn $
}

