% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.regexp.in.R
\name{SeaBird.regexp.in}
\alias{SeaBird.regexp.in}
\title{$Id: SeaBird.regexp.in.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.regexp.in(vector, regexp)
}
\description{
$Id: SeaBird.regexp.in.R 1747 2007-11-12 21:01:38Z adunn $
}
