% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.remove.last.words.R
\name{SeaBird.remove.last.words}
\alias{SeaBird.remove.last.words}
\title{$Id: SeaBird.remove.last.words.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.remove.last.words(string, words = 1)
}
\description{
$Id: SeaBird.remove.last.words.R 1747 2007-11-12 21:01:38Z adunn $
}

