% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.string.to.vector.of.numbers.R
\name{SeaBird.string.to.vector.of.numbers}
\alias{SeaBird.string.to.vector.of.numbers}
\title{$Id: SeaBird.string.to.vector.of.numbers.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.string.to.vector.of.numbers(string)
}
\description{
$Id: SeaBird.string.to.vector.of.numbers.R 1747 2007-11-12 21:01:38Z adunn $
}

