% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.string.to.vector.of.words.R
\name{SeaBird.string.to.vector.of.words}
\alias{SeaBird.string.to.vector.of.words}
\title{$Id: SeaBird.string.to.vector.of.words.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.string.to.vector.of.words(string)
}
\description{
$Id: SeaBird.string.to.vector.of.words.R 1747 2007-11-12 21:01:38Z adunn $
}
