% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.trim.lines.R
\name{SeaBird.trim.lines}
\alias{SeaBird.trim.lines}
\title{Trim rows so that any unfinished iterations are removed}
\usage{
SeaBird.trim.lines(lines)
}
\description{
Trim rows so that any unfinished iterations are removed
}
