% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.unpaste.R
\name{SeaBird.unpaste}
\alias{SeaBird.unpaste}
\title{$Id: SeaBird.unpaste.R 1747 2007-11-12 21:01:38Z adunn $}
\usage{
SeaBird.unpaste(string, sep)
}
\description{
$Id: SeaBird.unpaste.R 1747 2007-11-12 21:01:38Z adunn $
}
