% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeaBird.write.mark.recapture.R
\name{SeaBird.write_mark_recapture}
\alias{SeaBird.write_mark_recapture}
\title{Write mark recapture data block}
\usage{
SeaBird.write_mark_recapture(object, composite_labels, composite_map, file,
  path = "./", label, min.observations = 3)
}
\arguments{
\item{object}{\code{matrix} class object containing mark recapture data with labelled columns (years) and rows (band numbers).}

\item{composite_labels}{mark recapture data labels that correspond with composite classes}

\item{composite_map}{list of length equal to composite_labels with each entry containing a vector of labels assigned to each composite class}

\item{file}{output file name}

\item{path}{output file path}

\item{label}{label for data block}

\item{min.observations}{records of length <= min.observations are deleted}
}
\description{
Writes mark recapture data to a text file appropriate for insertion into \code{estimation.sbd}.
}
