% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.command.with.autonumber.R
\name{set.command.with.autonumber}
\alias{set.command.with.autonumber}
\title{set.command.with.autonumber}
\usage{
set.command.with.autonumber(sbd, command, autonumber, subcommand, argument)
}
\description{
set.command.with.autonumber
}

