% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.command.with.label.R
\name{set.command.with.label}
\alias{set.command.with.label}
\title{set.command.with.label}
\usage{
set.command.with.label(sbd, command, label, subcommand, argument)
}
\description{
set.command.with.label
}

