% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.sbd.file.R
\name{write.sbd.file}
\alias{write.sbd.file}
\title{$Id: write.sbd.file.R 1828 2007-11-30 01:16:07Z adunn $}
\usage{
write.sbd.file(object, file, path)
}
\description{
$Id: write.sbd.file.R 1828 2007-11-30 01:16:07Z adunn $
}

