% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeaBird.write_blocks.R
\name{write_blocks}
\alias{write_blocks}
\alias{write_blocks.tmatrix}
\title{Write transition matrix to text file}
\usage{
write_blocks(x, ...)

\method{write_blocks}{tmatrix}(object, path = ".")
}
\arguments{
\item{object}{\code{tmatrix} class object}
}
\description{
Two files are produced from the transition matrix input, called \code{transition_block.sbd} and \code{classes_block.sbd}. These should be pasted into \code{population.sbd}.
}

