% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.quantities.total_survival_at_class.R
\name{extract.quantities.total_survival_at_class}
\alias{extract.quantities.total_survival_at_class}
\title{extract.quantities.total_survival_at_class}
\usage{
extract.quantities.total_survival_at_class(lines)
}
\description{
extract.quantities.total_survival_at_class
}
