% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbd.commands.R
\docType{data}
\name{sbd.commands}
\alias{sbd.commands}
\title{$Id: sbd.commands.R 1750 2007-11-13 00:45:01Z adunn $}
\format{An object of class \code{list} of length 2.}
\usage{
sbd.commands
}
\description{
$Id: sbd.commands.R 1750 2007-11-13 00:45:01Z adunn $
}
\keyword{datasets}

